Run delete error files on prod only

+++ ingestion.js
@@ -2,7 +2,7 @@
 
 const yenv = require('yenv')
 const env = yenv('./env.yml', {
-	env: 'production'
+	env: 'default'
 })
 const aws = require('aws-sdk')
 const s3 = new aws.S3({
@@ -28,48 +28,53 @@ const advito = require('knex')({
 })
 
 module.exports.deleteErrorFiles = async () => {
-	const date = new Date()
-	date.setDate(date.getDate() - 5)
-	try {
-		console.log('Starting for this date: ', new Date())
-		const jobIngestionList = await advito('job_ingestion')
-			.select('id', 'file_name')
-			.andWhere('job_status', 'error')
-			.andWhere('upload_timestamp', '<=', date)
+	const Bucket = process.env.BUCKET ? process.env.BUCKET : env.BUCKET
+	if (Bucket === 'advito-pci') {
+		const date = new Date()
+		date.setDate(date.getDate() - 5)
+		try {
+			console.log('Starting for this date: ', new Date())
+			const jobIngestionList = await advito('job_ingestion')
+				.select('id', 'file_name')
+				.andWhere('job_status', 'error')
+				.andWhere('upload_timestamp', '<=', date)
 
-		console.log('jobIngestionList: ', jobIngestionList)
+			console.log('jobIngestionList: ', jobIngestionList)
 
-		const deleteArray = jobIngestionList
-			.filter((v) => v.file_name)
-			.map((v) => ({ Key: v.file_name }))
+			const deleteArray = jobIngestionList
+				.filter((v) => v.file_name)
+				.map((v) => ({ Key: v.file_name }))
 
-		if (deleteArray.length) {
-			const params = {
-				Bucket: process.env.BUCKET ? process.env.BUCKET : env.BUCKET,
-				Delete: {
-					Objects: deleteArray,
-					Quiet: false
+			if (deleteArray.length) {
+				const params = {
+					Bucket: process.env.BUCKET ? process.env.BUCKET : env.BUCKET,
+					Delete: {
+						Objects: deleteArray,
+						Quiet: false
+					}
 				}
+
+				console.log(`Deleting ${deleteArray.length} files.`)
+
+				await s3.deleteObjects(params).promise()
+
+				console.log('Updating list: ', jobIngestionList.length)
+				const updateList = jobIngestionList.map((job) =>
+					advito('job_ingestion')
+						.where('id', job.id)
+						.update({
+							job_status: 'deleted',
+							job_note: `Template Removed: ${new Date().toLocaleString()}`
+						})
+				)
+				await Promise.all(updateList)
 			}
-
-			console.log(`Deleting ${deleteArray.length} files.`)
-
-			await s3.deleteObjects(params).promise()
-
-			console.log('Updating list: ', jobIngestionList.length)
-			const updateList = jobIngestionList.map((job) =>
-				advito('job_ingestion')
-					.where('id', job.id)
-					.update({
-						job_status: 'deleted',
-						job_note: `Template Removed: ${new Date().toLocaleString()}`
-					})
-			)
-			await Promise.all(updateList)
+		} catch (e) {
+			console.log(`Error: ${e.message}`)
 		}
-	} catch (e) {
-		console.log(`Error: ${e.message}`)
+	} else {
+		console.log('This only works for production environment.')
 	}
 }
 
-module.exports.deleteErrorFiles()
+// module.exports.deleteErrorFiles()
